I #1525: Separated conversion built-ins tests
