+++ src/psyclone/tests/domain/lfric/lfric_builtins_test.py
@@ -4009,122 +4009,6 @@ def test_inc_min_aX(tmpdir, monkeypatch, annexed, dist_mem, fortran_writer):
         assert output_dm_2 in code
 
 
-# ------------- Converting real to integer field elements ------------------- #
-
-
-def test_int_X(tmpdir, monkeypatch, annexed, dist_mem):
-    '''Test that 1) the '__str__' method of 'LFRicIntXKern' returns the
-    expected string and 2) we generate correct code for the built-in
-    operation 'Y = int(X, i_def)' where 'Y' is an integer-valued
-    field, 'X' is the real-valued field being converted and the
-    correct kind, 'i_def', is picked up from the associated
-    field. Test with and without annexed DoFs being computed as this
-    affects the generated code. 3) Also test the 'metadata()' method.
-
-    '''
-    metadata = lfric_builtins.LFRicIntXKern.metadata()
-    assert isinstance(metadata, LFRicKernelMetadata)
-    api_config = Config.get().api_conf(API)
-    monkeypatch.setattr(api_config, "_compute_annexed_dofs", annexed)
-    _, invoke_info = parse(os.path.join(BASE_PATH,
-                                        "15.10.3_int_X_builtin.f90"),
-                           api=API)
-    psy = PSyFactory(API, distributed_memory=dist_mem).create(invoke_info)
-    # Test '__str__' method
-    first_invoke = psy.invokes.invoke_list[0]
-    kern = first_invoke.schedule.children[0].loop_body[0]
-    assert str(kern) == ("Built-in: int_X (convert a real-valued to an "
-                         "integer-valued field)")
-    # Test code generation
-    code = str(psy.gen)
-
-    assert LFRicBuild(tmpdir).code_compiles(psy)
-
-    # First check that the correct field types and constants are used
-    output = (
-        "    USE constants_mod, ONLY: r_def, i_def\n"
-        "    USE field_mod, ONLY: field_type, field_proxy_type\n"
-        "    USE integer_field_mod, ONLY: integer_field_type, "
-        "integer_field_proxy_type\n")
-    assert output in code
-
-    if not dist_mem:
-        output = (
-            "      INTEGER(KIND=i_def), pointer, dimension(:) :: f2_data => "
-            "null()\n"
-            "      TYPE(integer_field_proxy_type) f2_proxy\n"
-            "      REAL(KIND=r_def), pointer, dimension(:) :: f1_data => "
-            "null()\n"
-            "      TYPE(field_proxy_type) f1_proxy\n"
-            "      INTEGER(KIND=i_def) undf_aspc1_f2\n"
-            "      !\n"
-            "      ! Initialise field and/or operator proxies\n"
-            "      !\n"
-            "      f2_proxy = f2%get_proxy()\n"
-            "      f2_data => f2_proxy%data\n"
-            "      f1_proxy = f1%get_proxy()\n"
-            "      f1_data => f1_proxy%data\n"
-            "      !\n"
-            "      ! Initialise number of DoFs for aspc1_f2\n"
-            "      !\n"
-            "      undf_aspc1_f2 = f2_proxy%vspace%get_undf()\n"
-            "      !\n"
-            "      ! Set-up all of the loop bounds\n"
-            "      !\n"
-            "      loop0_start = 1\n"
-            "      loop0_stop = undf_aspc1_f2\n"
-            "      !\n"
-            "      ! Call our kernels\n"
-            "      !\n"
-            "      DO df=loop0_start,loop0_stop\n"
-            "        f2_data(df) = int(f1_data(df), i_def)\n"
-            "      END DO\n"
-            "      !\n"
-            "    END SUBROUTINE invoke_0\n")
-        assert output in code
-    else:
-        output_dm_2 = (
-            "      loop0_stop = f2_proxy%vspace%get_last_dof_annexed()\n"
-            "      !\n"
-            "      ! Call kernels and communication routines\n"
-            "      !\n"
-            "      DO df=loop0_start,loop0_stop\n"
-            "        f2_data(df) = int(f1_data(df), i_def)\n"
-            "      END DO\n"
-            "      !\n"
-            "      ! Set halos dirty/clean for fields modified in the "
-            "above loop\n"
-            "      !\n"
-            "      CALL f2_proxy%set_dirty()\n"
-            "      !\n")
-        if not annexed:
-            output_dm_2 = output_dm_2.replace("dof_annexed", "dof_owned")
-        assert output_dm_2 in code
-
-
-def test_int_X_precision(monkeypatch):
-    '''Test that the built-in picks up and creates correct code for a
-    scalar with precision that is not the default i.e. not 'i_def'. At
-    the moment there is no other integer precision so we make one up
-    and use monkeypatch to get round any error checks. However, this
-    does mean that we can't check whether it compiles. 3) Also test
-    the 'metadata()' method.
-
-    '''
-    metadata = lfric_builtins.LFRicIncATimesXKern.metadata()
-    assert isinstance(metadata, LFRicKernelMetadata)
-    _, invoke_info = parse(os.path.join(BASE_PATH,
-                                        "15.10.3_int_X_builtin.f90"),
-                           api=API)
-    psy = PSyFactory(API).create(invoke_info)
-    # Test '__str__' method
-    first_invoke = psy.invokes.invoke_list[0]
-    kern = first_invoke.schedule.children[0].loop_body[0]
-    monkeypatch.setattr(kern.args[0], "_precision", "i_solver")
-    code = str(psy.gen)
-    assert "USE constants_mod, ONLY: r_def, i_solver, i_def" in code
-    assert "f2_data(df) = int(f1_data(df), i_solver)" in code
-
 # ------------- Xfail built-ins --------------------------------------------- #
